requirements_prod.txt
fastapi==0.111.0
boto3==1.34.0
numpy==1.26.4
onnxruntime==1.16.3
pillow==10.3.0
python-multipart==0.0.9
mlflow-skinny==2.14.1
mangum==0.17.0
